Rename files related to class URL to avoid collisions
 ... on case insensative filesystems with acl/Url.*
 when building the src/acl/ objects.

Rename:
 src/url.cc -> src/anyp/Uri.cc
 src/URL.h -> src/anyp/Uri.h
 src/tests/stub_url.cc -> src/tests/stub_libanyp.cc

Rational for using "Uri" is that this class actually
stores any URI (both URL and URN) whether complete
or not.

Also, remove many unnecessary #include statements.
Since any/Uri.h is pulled in through HttpRequest.h
and HttpReply.h.
