+++ src/Makefile.am
@@ -476,8 +476,6 @@ squid_SOURCES = \
 	tunnel.cc \
 	typedefs.h \
 	$(UNLINKDSOURCE) \
-	url.cc \
-	URL.h \
 	urn.h \
 	urn.cc \
 	wccp.h \
@@ -996,7 +994,7 @@ tests_testHttpReply_SOURCES=\
 	tests/testHttpReply.cc \
 	tests/testHttpReply.h \
 	tests/stub_time.cc \
-	tests/stub_url.cc \
+	tests/stub_libanyp.cc \
 	wordlist.h \
 	wordlist.cc
 nodist_tests_testHttpReply_SOURCES=\
@@ -1118,8 +1116,7 @@ tests_testACLMaxUserIP_SOURCES= \
 	tests/testACLMaxUserIP.cc \
 	tests/testACLMaxUserIP.h \
 	tests/stub_time.cc \
-	tests/stub_url.cc \
-	URL.h \
+	tests/stub_libanyp.cc \
 	MemBuf.cc \
 	wordlist.h \
 	wordlist.cc
@@ -1369,7 +1366,7 @@ tests_testCacheManager_SOURCES = \
 	tests/stub_SwapDir.cc \
 	MemStore.cc \
 	$(UNLINKDSOURCE) \
-	tests/stub_url.cc \
+	tests/stub_libanyp.cc \
 	urn.h \
 	urn.cc \
 	wccp2.h \
@@ -1551,7 +1548,7 @@ tests_testDiskIO_SOURCES = \
 	tests/testStoreSupport.h \
 	tests/stub_time.cc \
 	$(UNLINKDSOURCE) \
-	tests/stub_url.cc \
+	tests/stub_libanyp.cc \
 	$(WIN32_SOURCE) \
 	wordlist.h \
 	wordlist.cc \
@@ -1800,7 +1797,7 @@ tests_testEvent_SOURCES = \
 	tests/stub_tunnel.cc \
 	MemStore.cc \
 	$(UNLINKDSOURCE) \
-	tests/stub_url.cc \
+	tests/stub_libanyp.cc \
 	urn.h \
 	urn.cc \
 	wccp2.h \
@@ -2034,7 +2031,7 @@ tests_testEventLoop_SOURCES = \
 	tests/stub_tunnel.cc \
 	MemStore.cc \
 	$(UNLINKDSOURCE) \
-	tests/stub_url.cc \
+	tests/stub_libanyp.cc \
 	urn.h \
 	urn.cc \
 	wccp2.h \
@@ -2263,7 +2260,7 @@ tests_test_http_range_SOURCES = \
 	tools.cc \
 	tests/stub_tunnel.cc \
 	$(UNLINKDSOURCE) \
-	tests/stub_url.cc \
+	tests/stub_libanyp.cc \
 	urn.h \
 	urn.cc \
 	wccp2.h \
@@ -2368,7 +2365,7 @@ tests_testHttp1Parser_SOURCES = \
 	tests/testHttp1Parser.cc \
 	tests/testHttp1Parser.h \
 	tests/stub_time.cc \
-	tests/stub_url.cc \
+	tests/stub_libanyp.cc \
 	wordlist.h \
 	wordlist.cc
 nodist_tests_testHttp1Parser_SOURCES = \
@@ -2565,7 +2562,6 @@ tests_testHttpRequest_SOURCES = \
 	tests/stub_tunnel.cc \
 	tests/stub_SwapDir.cc \
 	MemStore.cc \
-	url.cc \
 	urn.h \
 	urn.cc \
 	wccp2.h \
@@ -2801,7 +2797,7 @@ tests_testStore_SOURCES= \
 	tests/TestSwapDir.cc \
 	tests/TestSwapDir.h \
 	tests/stub_time.cc \
-	tests/stub_url.cc \
+	tests/stub_libanyp.cc \
 	wordlist.h \
 	wordlist.cc
 
@@ -2998,7 +2994,7 @@ tests_testUfs_SOURCES = \
 	HttpHdrCc.cc \
 	HttpHdrSc.cc \
 	HttpHdrScTarget.cc \
-	tests/stub_url.cc \
+	tests/stub_libanyp.cc \
 	StatCounters.h \
 	StatCounters.cc \
 	StatHist.h \
@@ -3198,7 +3194,7 @@ tests_testRock_SOURCES = \
 	tools.h \
 	tests/stub_tools.cc \
 	time.cc \
-	tests/stub_url.cc \
+	tests/stub_libanyp.cc \
 	wordlist.h \
 	wordlist.cc \
 	$(DELAY_POOL_SOURCE) \
@@ -3424,7 +3420,6 @@ tests_testURL_SOURCES = \
 	tools.h \
 	tools.cc \
 	tests/stub_tunnel.cc \
-	url.cc \
 	urn.h \
 	urn.cc \
 	wccp2.h \
