+++ src/acl/Url.cc
@@ -14,7 +14,6 @@
 #include "acl/Url.h"
 #include "HttpRequest.h"
 #include "rfc1738.h"
-#include "src/URL.h"
 
 int
 ACLUrlStrategy::match (ACLData<char const *> * &data, ACLFilledChecklist *checklist)
