+++ src/tests/Stub.list
@@ -41,6 +41,7 @@ STUB_SOURCE= tests/STUB.h \
 	tests/stub_ipc.cc \
 	tests/stub_ipc_Forwarder.cc \
 	tests/stub_ipc_TypedMsgHdr.cc \
+	tests/stub_libanyp.cc \
 	tests/stub_libauth_acls.cc \
 	tests/stub_libauth.cc \
 	tests/stub_libcomm.cc \
@@ -81,7 +82,6 @@ STUB_SOURCE= tests/STUB.h \
 	tests/stub_tools.cc \
 	tests/stub_tunnel.cc \
 	tests/stub_UdsOp.cc \
-	tests/stub_url.cc \
 	tests/stub_wccp2.cc \
 	tests/stub_whois.cc \
 	tests/stub_wordlist.cc
